ANDROID: binder: Fix binder_transaction_priority call when TF_ONE_WAY is not set
This is a quick workaround the "ANDROID: binder: improve priority inheritance."
backport where the current backport requires that the target_thread is always
non-NULL which is not the case.

I don't understand the full extent of the android binder IPC but the upstream
commit relies on a per process queue (of threads) which this version of binder
obviously does not have. The binder_proc_transaction() on the upstream commit
would try to find a thread which will be woken up or will use the thread
parameter if it's non-NULL.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>