arm: mach-sc: Do not use a spinlock on the ISR
The spinlock appears to guard the whole routine from simultaneous execution
from other CPUs or cores even though the kernel already guarantees that
an ISR or an interrupt handler will only run on a single cpu at any time.
This notion is stated in the "Unreliable Guide To Locking" document under the
"Data Which Mostly Used By An IRQ Handler" section.

On the other side of this change, the lock and unlock operations takes 10-20
usec meaning that we could shave off about 20 usec of time running this
time sensitive ISR.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>